backend/requirements.txt
flask==2.2.5
flask_cors==5.0.0
flask-socketio==5.3.0
Werkzeug==3.1.3
pymongo[srv]==3.11
gevent-websocket